kola/tests/iso: refactor live-iso.go into independent test files
Split the monolithic live-iso.go file into several focused files where
each file contains a specific set of tests:

- live-pxe.go: PXE network boot tests
- live-iscsi.go: iSCSI boot and installation tests
- live-as-disk.go: ISO-as-disk installation tests
- common.go: Shared test utilities and helper functions

This improves code organization and makes it easier to locate and
maintain specific test suites.
