+++ mantle/kola/tests/iso/live-login.go
@@ -1,4 +1,4 @@
-package testiso
+package iso
 
 import (
 	"bufio"
