+++ mantle/kola/tests/iso/live-pxe.go
@@ -1,0 +1,172 @@
+package iso
+
+import (
+	"fmt"
+	"os"
+	"strings"
+	"time"
+
+	"github.com/coreos/coreos-assembler/mantle/kola"
+	"github.com/coreos/coreos-assembler/mantle/kola/cluster"
+	"github.com/coreos/coreos-assembler/mantle/kola/register"
+	"github.com/coreos/coreos-assembler/mantle/platform/conf"
+	"github.com/coreos/coreos-assembler/mantle/platform/machine/qemu"
+	"github.com/coreos/coreos-assembler/mantle/util"
+	coreosarch "github.com/coreos/stream-metadata-go/arch"
+)
+
+var (
+	tests_pxe_x86_64 = []string{
+		"pxe-offline-install.rootfs-appended.bios",
+		"pxe-offline-install.4k.uefi",
+		"pxe-online-install.bios",
+		"pxe-online-install.4k.uefi",
+	}
+	tests_pxe_aarch64 = []string{
+		"pxe-offline-install.uefi",
+		"pxe-offline-install.rootfs-appended.4k.uefi",
+		"pxe-online-install.uefi",
+		"pxe-online-install.4k.uefi",
+	}
+	tests_pxe_ppc64le = []string{
+		"pxe-online-install.rootfs-appended.ppcfw",
+		"pxe-offline-install.4k.ppcfw",
+	}
+	tests_pxe_s390x = []string{
+		"pxe-online-install.rootfs-appended.s390fw",
+		"pxe-offline-install.s390fw",
+	}
+)
+
+func getAllPxeTests() []string {
+	arch := coreosarch.CurrentRpmArch()
+	switch arch {
+	case "x86_64":
+		return tests_pxe_x86_64
+	case "aarch64":
+		return tests_pxe_aarch64
+	case "ppc64le":
+		return tests_pxe_ppc64le
+	case "s390x":
+		return tests_pxe_s390x
+	default:
+		return []string{}
+	}
+}
+
+func init() {
+	for _, testName := range getAllPxeTests() {
+		register.RegisterTest(&register.Test{
+			Run: func(c cluster.TestCluster) {
+				opts := getIsoTestOpts(testName)
+				testPXE(c, opts)
+			},
+			ClusterSize: 0,
+			Name:        "iso." + testName,
+			Description: "Verify PXE install works.",
+			Timeout:     installTimeoutMins * time.Minute,
+			Flags:       []register.Flag{},
+			Platforms:   []string{"qemu"},
+		})
+	}
+}
+
+var downloadCheck = `[Unit]
+Description=TestISO Verify CoreOS Installer Download
+After=coreos-installer.service
+Before=coreos-installer.target
+[Service]
+Type=oneshot
+StandardOutput=kmsg+console
+StandardError=kmsg+console
+ExecStart=/bin/sh -c "journalctl -t coreos-installer-service | /usr/bin/awk '/[Dd]ownload/ {exit 1}'"
+ExecStart=/bin/sh -c "/usr/bin/udevadm settle"
+ExecStart=/bin/sh -c "/usr/bin/mount /dev/disk/by-label/root /mnt"
+ExecStart=/bin/sh -c "/usr/bin/jq -er '.[\"build\"]? + .[\"version\"]? == \"%s\"' /mnt/.coreos-aleph-version.json"
+[Install]
+RequiredBy=coreos-installer.target
+`
+
+func testPXE(c cluster.TestCluster, opts IsoTestOpts) {
+	var outdir string
+	var qc *qemu.Cluster
+
+	switch pc := c.Cluster.(type) {
+	case *qemu.Cluster:
+		outdir = pc.RuntimeConf().OutputDir
+		qc = pc
+	default:
+		c.Fatalf("Unsupported cluster type")
+	}
+
+	if opts.addNmKeyfile {
+		c.Fatal("--add-nm-keyfile not yet supported for PXE")
+	}
+
+	inst := qemu.Install{
+		CosaBuild:       kola.CosaBuild,
+		NmKeyfiles:      make(map[string]string),
+		Insecure:        opts.instInsecure,
+		Native4k:        opts.enable4k,
+		MultiPathDisk:   opts.enableMultipath,
+		PxeAppendRootfs: opts.pxeAppendRootfs,
+	}
+
+	tmpd, err := os.MkdirTemp("", "kola-iso.pxe")
+	if err != nil {
+		c.Fatal(err)
+	}
+	defer os.RemoveAll(tmpd)
+
+	sshPubKeyBuf, _, err := util.CreateSSHAuthorizedKey(tmpd)
+	if err != nil {
+		c.Fatal(err)
+	}
+
+	builder, virtioJournalConfig, err := newQemuBuilderWithDisk(opts, outdir)
+	if err != nil {
+		c.Fatal(err)
+	}
+
+	// increase the memory for pxe tests with appended rootfs in the initrd
+	// we were bumping up into the 4GiB limit in RHCOS/c9s
+	// pxe-offline-install.rootfs-appended.bios tests
+	if inst.PxeAppendRootfs && builder.MemoryMiB < 5120 {
+		builder.MemoryMiB = 5120
+	}
+
+	inst.Builder = builder
+	completionChannel, err := inst.Builder.VirtioChannelRead("testisocompletion")
+	if err != nil {
+		c.Fatal(err) // , "setting up virtio-serial channel")
+	}
+
+	var keys []string
+	keys = append(keys, strings.TrimSpace(string(sshPubKeyBuf)))
+	virtioJournalConfig.AddAuthorizedKeys("core", keys)
+
+	liveConfig := *virtioJournalConfig
+	liveConfig.AddSystemdUnit("live-signal-ok.service", liveSignalOKUnit, conf.Enable)
+	liveConfig.AddSystemdUnit("coreos-test-entered-emergency-target.service", signalFailureUnit, conf.Enable)
+
+	if opts.isOffline {
+		contents := fmt.Sprintf(downloadCheck, kola.CosaBuild.Meta.OstreeVersion)
+		liveConfig.AddSystemdUnit("coreos-installer-offline-check.service", contents, conf.Enable)
+	}
+
+	targetConfig := *virtioJournalConfig
+	targetConfig.AddSystemdUnit("coreos-test-installer.service", signalCompletionUnit, conf.Enable)
+	targetConfig.AddSystemdUnit("coreos-test-entered-emergency-target.service", signalFailureUnit, conf.Enable)
+	targetConfig.AddSystemdUnit("coreos-test-installer-no-ignition.service", checkNoIgnition, conf.Enable)
+
+	mach, err := inst.PXE(opts.pxeKernelArgs, liveConfig, targetConfig, opts.isOffline)
+	if err != nil {
+		c.Fatal(err)
+	}
+	qc.AddMach(mach)
+
+	err = awaitCompletion(c, mach.Instance(), opts.console, outdir, completionChannel, mach.BootStartedErrorChannel(), []string{liveOKSignal, signalCompleteString})
+	if err != nil {
+		c.Fatal(err)
+	}
+}
